+++ bkend/common/src/main/java/com/sourcekraft/documentburster/common/settings/model/FreeMarkerSettings.java
@@ -19,4 +19,8 @@ public class FreeMarkerSettings extends DumpToString {
     
     // Number format (optional)
     public String numberformat;
+    
+    // Timezone
+    public String timezone;
+    
 }